+++ src/Components/Admins/List.js
@@ -3,30 +3,36 @@ import styles from './admins.module.css';
 import Table from './Table';
 import Button from './Button';
 import Form from './Form';
+import UpdateForm from './UpdateForm';
 
-const List = ({ admins, addAdmin, deleteAdmin, adminToUpdate, updateAdmin }) => {
+const List = ({ admins, addAdmin, deleteAdmin, updateAdmin }) => {
   const [showForm, setForm] = useState(false);
+  const [showEditForm, setEditForm] = useState(false);
+  const [adminToUpdate, setId] = useState('');
+
+  const setupEditForm = (admin = '') => {
+    setId(admin);
+    setEditForm(!showEditForm);
+  };
 
   return (
     <section className={styles.list}>
       <header className={styles.header}>
         <h1 className={styles.title}>Administrators</h1>
         <Button
-          color={showForm ? '#FF3E1D' : '#46A37C'}
+          setStyle={showForm ? `${styles.btnClosed}` : `${styles.btn}`}
           text={showForm ? 'Close' : 'Add Admin'}
           onclick={() => setForm(!showForm)}
         />
       </header>
       {showForm && <Form addAdmin={addAdmin} />}
       {admins.length > 0 ? (
-        <Table
-          admins={admins}
-          deleteAdmin={deleteAdmin}
-          adminToUpdate={adminToUpdate}
-          updateAdmin={updateAdmin}
-        />
+        <Table admins={admins} deleteAdmin={deleteAdmin} onEdit={setupEditForm} />
       ) : (
         <span className={styles.emptySpan}>No admins created yet</span>
+      )}
+      {showEditForm && (
+        <UpdateForm close={setupEditForm} updateAdmin={updateAdmin} admin={adminToUpdate} />
       )}
     </section>
   );
