+++ src/Components/Admins/UpdateForm.js
@@ -31,94 +31,89 @@ const UpdateForm = ({ close, admin, updateAdmin }) => {
   };
 
   return (
-    <div className={styles.modal}>
-      <div className={styles.modalContent}>
-        <div className={styles.modalHeader}>
-          <label>Update Admin</label>
-          <span onClick={close} style={{ cursor: 'pointer' }}>
-            &times;
-          </span>
+    <div className={styles.modalContent}>
+      <div className={styles.modalHeader}>
+        <h3>Update Admin</h3>
+      </div>
+      <form className={styles.modalForm}>
+        <div className={styles.inputGroup}>
+          <label>First Name</label>
+          <input
+            type="text"
+            placeholder="First Name"
+            name="firstName"
+            value={adminUpdated.firstName}
+            onChange={changeHandler}
+          />
         </div>
-        <form className={styles.modalForm}>
-          <div className={styles.inputGroup}>
-            <label>First Name</label>
-            <input
-              type="text"
-              placeholder="First Name"
-              name="firstName"
-              value={adminUpdated.firstName}
-              onChange={changeHandler}
-            />
-          </div>
-          <div className={styles.inputGroup}>
-            <label>Last Name</label>
-            <input
-              type="text"
-              placeholder="Last Name"
-              name="lastName"
-              value={adminUpdated.lastName}
-              onChange={changeHandler}
-            />
-          </div>
-          <div className={styles.inputGroup}>
-            <label>DNI</label>
-            <input
-              type="number"
-              placeholder="DNI number"
-              name="dni"
-              value={adminUpdated.dni}
-              onChange={changeHandler}
-            />
-          </div>
-          <div className={styles.inputGroup}>
-            <label>Phone</label>
-            <input
-              type="number"
-              placeholder="Phone number"
-              name="phone"
-              value={adminUpdated.phone}
-              onChange={changeHandler}
-            />
-          </div>
-          <div className={styles.inputGroup}>
-            <label>Email</label>
-            <input
-              type="text"
-              placeholder="Email"
-              name="email"
-              value={adminUpdated.email}
-              onChange={changeHandler}
-            />
-          </div>
-          <div className={styles.inputGroup}>
-            <label>City</label>
-            <input
-              type="text"
-              placeholder="City"
-              name="city"
-              value={adminUpdated.city}
-              onChange={changeHandler}
-            />
-          </div>
-          <div className={styles.inputGroup}>
-            <label>Password</label>
-            <input
-              type="password"
-              placeholder="Password"
-              name="password"
-              value={adminUpdated.password}
-              onChange={changeHandler}
-            />
-          </div>
-        </form>
-        <div className={styles.modalBtns}>
-          <button className={styles.saveBtn} onClick={saveUpdate}>
-            Save
-          </button>
-          <button className={styles.closeBtn} onClick={cancelUpdate}>
-            Cancel
-          </button>
+        <div className={styles.inputGroup}>
+          <label>Last Name</label>
+          <input
+            type="text"
+            placeholder="Last Name"
+            name="lastName"
+            value={adminUpdated.lastName}
+            onChange={changeHandler}
+          />
         </div>
+        <div className={styles.inputGroup}>
+          <label>DNI</label>
+          <input
+            type="text"
+            placeholder="DNI number"
+            name="dni"
+            value={adminUpdated.dni}
+            onChange={changeHandler}
+          />
+        </div>
+        <div className={styles.inputGroup}>
+          <label>Phone</label>
+          <input
+            type="text"
+            placeholder="Phone number"
+            name="phone"
+            value={adminUpdated.phone}
+            onChange={changeHandler}
+          />
+        </div>
+        <div className={styles.inputGroup}>
+          <label>Email</label>
+          <input
+            type="text"
+            placeholder="Email"
+            name="email"
+            value={adminUpdated.email}
+            onChange={changeHandler}
+          />
+        </div>
+        <div className={styles.inputGroup}>
+          <label>City</label>
+          <input
+            type="text"
+            placeholder="City"
+            name="city"
+            value={adminUpdated.city}
+            onChange={changeHandler}
+          />
+        </div>
+        <div className={styles.inputGroup}>
+          <label>Password</label>
+          <input
+            type="password"
+            placeholder="Password"
+            name="password"
+            value={adminUpdated.password}
+            onChange={changeHandler}
+          />
+        </div>
+      </form>
+      <div className={styles.modalBtns}>
+        <button className={styles.closeBtn} onClick={cancelUpdate}>
+          Cancel
+        </button>
+        <button className={styles.saveBtn} onClick={saveUpdate}>
+          Save
+        </button>
       </div>
     </div>
   );
