+++ src/Components/Admins/admins.module.css
@@ -62,6 +62,20 @@ td {
   cursor: pointer;
 }
 
+.btnClosed {
+  width: 150px;
+  height: 50px;
+  background-color: #FF3E1D;
+  border: none;
+  border-radius: 5px;
+  font-size: 16px;
+  font-weight: bold;
+  line-height: 19px;
+  color: white;
+  text-transform: uppercase;
+  cursor: pointer;
+}
+
 .table {
   width: 80%;
   margin-top: 20px;
@@ -152,24 +166,12 @@ td {
   font-weight: bold;
 }
 
-.modal {
-  position: absolute;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0,0,0,0.4  );
-}
-
 .modalContent {
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin: 15% auto;
-  width: 700px;
-  height: 550px;
-  background-color: #2F3058;
+  width: 80%;
+  margin-top: 20px;
   border-radius: 5px;
 }
 
@@ -177,7 +179,7 @@ td {
   display: flex;
   width: 100%;
   justify-content: space-between;
-  padding: 20px;
+  padding: 20px 0;
   color: white;
   font-size: 25px;
   font-weight: bold;
@@ -186,13 +188,12 @@ td {
 .modalForm {
   display: flex;
   flex-wrap: wrap;
-  justify-content: center;
-  padding: 20px;
 }
 
 .modalBtns {
   display: flex;
   width: 50%;
+  margin-top: 20px;
   justify-content: space-around;
 }
